issue #123 : Extract DownloadControlThread to a separate class
The core domain specific content in DownloadControlThread was the creation of callables, which has been extracted using a new functional interface named DownloadCreator.

This will enable reuse of the components/scheduling tools created here in other parts of biocache.
